feat: extract trace level into separate header file
